# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Property Portal API",
        "description": "API documentation for the Property Portal backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "List all active properties with filters",
                "operationId": "e83a7fc5b1d98569892c10b7b0223879",
                "parameters": [
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "Destination slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "State slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "collection",
                        "in": "query",
                        "description": "Collection slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "check_in",
                        "in": "query",
                        "description": "Format: YYYY-MM-DD",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "check_out",
                        "in": "query",
                        "description": "Format: YYYY-MM-DD",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "guests",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rooms_requested",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Radius in km (default 50)",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "amenities",
                        "in": "query",
                        "description": "Comma-separated amenity IDs e.g. 1,3,5",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of properties with full details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/{slug}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get single property details by slug",
                "operationId": "3b9a5917561a9d6e3c3260bfff324ed5",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the property",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed property information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/{slug}/related": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get related properties by slug",
                "operationId": "e8449de6a3cbe2cba9776b57409d8255",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the property",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of related properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}